requirements-rpi.txt
socketIO-client
RPi.GPIO
